Add some coordinates function and an implementation for W
This may not be the approach that I want to continue with, but for now
add a coords{x, y} struct and some basic operations to
perform (eg. equality checks).

get_coords_W just returns a point now, rather than mutating the editor
state.

This should all be refactored soon.